This change speeds up rendering of any page shown inside the community layout by roughly 2X.
Explanation: FieldProperty objects are data descriptors, and are therefore 
invoked on every attr lookup (e.g. "tag.name").  If an attribute does
not exist on the instance, the FieldProperty is willing to provide a default
value from the ITag interface.  But since every attr is present by default
anyway, this is a useless wheel-spin here.

Removing this feature removes 40K (!) function calls from every page
rendering that uses the community layout.